+++ clang/lib/Format/FormatTokenLexer.cpp
@@ -785,7 +785,7 @@ bool FormatTokenLexer::tryMergeConflictMarkers() {
 
   FileID ID;
   // Get the position of the first token in the line.
-  unsigned FirstInLineOffset;
+  uint64_t FirstInLineOffset;
   std::tie(ID, FirstInLineOffset) = SourceMgr.getDecomposedLoc(
       Tokens[FirstInLineIndex]->getStartOfNonWhitespace());
   StringRef Buffer = SourceMgr.getBufferOrFake(ID).getBuffer();
